requirements.txt
opencv-python==4.10.0.84
pyzbar
Pillow
qrcode
imageio